requirements.txt
# Web处理相关
DrissionPage>=4.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# LLM交互相关
openai>=1.12.0
python-dotenv==1.0.0
httpx>=0.24.0

# 模板引擎
jinja2==3.1.2

# 图像处理
Pillow==10.1.0

# 数据处理
pydantic==2.5.3

# 日志和工具
loguru==0.7.2
tqdm==4.66.1

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

